Text change support for Class + Saga
