+++ forge-game/src/main/java/forge/game/TriggerReplacementBase.java
@@ -1,6 +1,7 @@
 package forge.game;
 
 import java.util.EnumSet;
+import java.util.Map;
 import java.util.Set;
 
 import forge.game.card.Card;
@@ -85,4 +86,35 @@ public abstract class TriggerReplacementBase extends CardTraitBase implements II
     }
 
     abstract public SpellAbility ensureAbility();
+
+    /* (non-Javadoc)
+     * @see forge.game.CardTraitBase#changeText()
+     */
+    @Override
+    public void changeText() {
+        if (!isIntrinsic()) {
+            return;
+        }
+        super.changeText();
+
+        SpellAbility sa = ensureAbility();
+
+        if (sa != null) {
+            sa.changeText();
+        }
+    }
+
+    /* (non-Javadoc)
+     * @see forge.game.CardTraitBase#changeTextIntrinsic(java.util.Map, java.util.Map)
+     */
+    @Override
+    public void changeTextIntrinsic(Map<String, String> colorMap, Map<String, String> typeMap) {
+        super.changeTextIntrinsic(colorMap, typeMap);
+
+        SpellAbility sa = ensureAbility();
+
+        if (sa != null) {
+            sa.changeTextIntrinsic(colorMap, typeMap);
+        }
+    }
 }
