+++ forge-game/src/main/java/forge/game/card/Card.java
@@ -5299,6 +5299,14 @@ public class Card extends GameEntity implements Comparable<Card>, IHasSVars {
                 KeywordInterface newKw = Keyword.getInstance(newtxt);
                 addKeywords.add(newKw);
                 removeKeywords.add(kw);
+            } else if (oldtxt.startsWith("Class")) {
+                for (StaticAbility trait : kw.getStaticAbilities()) {
+                    trait.changeText();
+                }
+            } else if (oldtxt.startsWith("Chapter")) {
+                for (Trigger trait : kw.getTriggers()) {
+                    trait.changeText();
+                }
             }
         }
