+++ forge-game/src/main/java/forge/game/trigger/Trigger.java
@@ -592,37 +592,6 @@ public abstract class Trigger extends TriggerReplacementBase {
         }
     }
 
-    /* (non-Javadoc)
-     * @see forge.game.CardTraitBase#changeText()
-     */
-    @Override
-    public void changeText() {
-        if (!isIntrinsic()) {
-            return;
-        }
-        super.changeText();
-
-        SpellAbility sa = ensureAbility();
-
-        if (sa != null) {
-            sa.changeText();
-        }
-    }
-
-    /* (non-Javadoc)
-     * @see forge.game.CardTraitBase#changeTextIntrinsic(java.util.Map, java.util.Map)
-     */
-    @Override
-    public void changeTextIntrinsic(Map<String, String> colorMap, Map<String, String> typeMap) {
-        super.changeTextIntrinsic(colorMap, typeMap);
-
-        SpellAbility sa = ensureAbility();
-
-        if (sa != null) {
-            sa.changeTextIntrinsic(colorMap, typeMap);
-        }
-    }
-
     public SpellAbility ensureAbility(final IHasSVars sVarHolder) {
         SpellAbility sa = getOverridingAbility();
         if (sa == null && hasParam("Execute")) {
